requirements.txt
# FastAPI and web server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Pydantic for data validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.0

# AWS S3
boto3>=1.34.0
botocore>=1.34.0

# Azure Blob Storage
azure-storage-blob>=12.19.0
azure-identity>=1.15.0

# Google Cloud Storage
google-cloud-storage>=2.14.0

# Utilities
python-dotenv>=1.0.0
python-dateutil>=2.8.0
pyyaml>=6.0

# Kafka
aiokafka>=0.10.0
kafka-python>=2.0.2

# Scheduler
apscheduler>=3.10.0

# Vector embeddings and search
sentence-transformers>=2.2.0
numpy>=1.24.0
scikit-learn>=1.3.0


# Background Tasks
apscheduler>=3.10.0
